+++ packages/math-inline/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.11.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/math-inline@7.10.0...@pie-element/math-inline@7.11.0) (2024-03-26)
+
+
+### Features
+
+* Replace math-rendering package with math-rendering-accessible-package for accessibility PD-1870 ([a416664](https://github.com/pie-framework/pie-elements/commit/a41666452b86426a35e8501ab2a4ac95f327ad53))
+
+
+
+
+
+# [7.10.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/math-inline@7.9.0...@pie-element/math-inline@7.10.0) (2024-03-26)
+
+
+### Features
+
+* extended the function that was searching for the data-keypad attribute PD-3593 ([b75c0e7](https://github.com/pie-framework/pie-elements/commit/b75c0e78bdce93db1ba5350b5c55572411e78eea))
+
+
+
+
+
 # [7.9.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/math-inline@7.8.0...@pie-element/math-inline@7.9.0) (2024-03-20)